Add Enzyme test
