+++ package.json
@@ -11,7 +11,7 @@
     "react-dev": "webpack -d --watch",
     "server-dev": "nodemon server/index.js",
     "seed": "mysql -u root -p < db/init.sql && node db/index.js",
-    "test": "jest"
+    "test": "jest --env=jsdom"
   },
   "dependencies": {
     "axios": "^0.18.0",
@@ -19,8 +19,10 @@
     "babel-preset-env": "^1.6.1",
     "babel-preset-es2015": "^6.24.1",
     "babel-preset-react": "^6.24.1",
+    "cors": "^2.8.4",
     "css-loader": "^1.0.1",
     "express": "^4.16.4",
+    "jsdom": "^13.0.0",
     "knex": "^0.15.2",
     "lorem-ipsum": "^1.0.6",
     "mysql": "^2.16.0",
@@ -37,6 +39,8 @@
     "babel-core": "^6.26.3",
     "babel-jest": "^23.6.0",
     "chai": "^4.2.0",
+    "enzyme": "^3.7.0",
+    "enzyme-adapter-react-16": "^1.6.0",
     "eslint-config-hackreactor": "git://github.com/reactorcore/eslint-config-hackreactor",
     "jest": "^23.6.0",
     "mocha": "^5.2.0",
