+++ server/index.js
@@ -2,10 +2,13 @@ const express = require('express');
 const morgan = require('morgan');
 const path = require('path');
 const models = require('./../db/models');
+var cors = require('cors')
 
 const app = express();
 
 app.use(morgan('dev'));
+
+app.use(cors())
 
 app.use(express.static(path.join(__dirname, './../client/dist')));
